refactor: Route loaders. Refactor `ApiLoader`.
