+++ src/Bridge/Symfony/Routing/Loader/ResourceLoader.php
@@ -9,16 +9,10 @@ use ApiPlatform\Core\Bridge\Symfony\Routing\RouteNameGenerator;
 use ApiPlatform\Core\Exception\InvalidResourceException;
 use ApiPlatform\Core\Exception\RuntimeException;
 use ApiPlatform\Core\Metadata\Resource\Factory\ResourceMetadataFactoryInterface;
-use ApiPlatform\Core\Metadata\Resource\Factory\ResourceNameCollectionFactoryInterface;
 use ApiPlatform\Core\Metadata\Resource\ResourceMetadata;
-use ApiPlatform\Core\Operation\Factory\SubresourceOperationFactoryInterface;
 use ApiPlatform\Core\PathResolver\OperationPathResolverInterface;
-use Symfony\Component\Config\FileLocator;
 use Symfony\Component\Config\Loader\Loader;
-use Symfony\Component\Config\Resource\DirectoryResource;
 use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\HttpKernel\KernelInterface;
-use Symfony\Component\Routing\Loader\XmlFileLoader;
 use Symfony\Component\Routing\Route;
 use Symfony\Component\Routing\RouteCollection;
 
@@ -26,6 +20,7 @@ use Symfony\Component\Routing\RouteCollection;
  * Loads Resources.
  *
  * @author Kévin Dunglas <dunglas@gmail.com>
+ * @author Tebaly <admin@freedomsex.net>
  */
 final class ResourceLoader extends Loader
 {
@@ -34,18 +29,15 @@ final class ResourceLoader extends Loader
     private $resourceMetadataFactory;
     private $operationPathResolver;
     private $container;
-    private $resourceClassDirectories;
 
     public function __construct(
         ResourceMetadataFactoryInterface $resourceMetadataFactory,
         OperationPathResolverInterface $operationPathResolver,
-        ContainerInterface $container,
-        array $resourceClassDirectories = []
+        ContainerInterface $container
     ) {
         $this->resourceMetadataFactory = $resourceMetadataFactory;
         $this->operationPathResolver = $operationPathResolver;
         $this->container = $container;
-        $this->resourceClassDirectories = $resourceClassDirectories;
     }
 
     /**
@@ -53,27 +45,28 @@ final class ResourceLoader extends Loader
      */
     public function load($resourceClass, $type = null): RouteCollection
     {
-        $routeCollection = new RouteCollection();
+        $collection = new RouteCollection();
         $resourceMetadata = $this->resourceMetadataFactory->create($resourceClass);
         $resourceShortName = $resourceMetadata->getShortName();
 
         if (null === $resourceShortName) {
             throw new InvalidResourceException(sprintf('Resource %s has no short name defined.', $resourceClass));
         }
-
         if (null !== $collectionOperations = $resourceMetadata->getCollectionOperations()) {
             foreach ($collectionOperations as $operationName => $operation) {
-                $this->addRoute($routeCollection, $resourceClass, $operationName, $operation, $resourceMetadata, OperationType::COLLECTION);
+                $this->addRoute($collection, $resourceClass, $operationName, $operation, $resourceMetadata, OperationType::COLLECTION);
             }
         }
-
         if (null !== $itemOperations = $resourceMetadata->getItemOperations()) {
             foreach ($itemOperations as $operationName => $operation) {
-                $this->addRoute($routeCollection, $resourceClass, $operationName, $operation, $resourceMetadata, OperationType::ITEM);
+                $this->addRoute($collection, $resourceClass, $operationName, $operation, $resourceMetadata, OperationType::ITEM);
             }
         }
 
-        return $routeCollection;
+        $subCollection = $this->import($resourceClass, 'api_subresource');
+        $collection->addCollection($subCollection);
+
+        return $collection;
     }
 
     /**
@@ -81,7 +74,6 @@ final class ResourceLoader extends Loader
      */
     public function supports($resource, $type = null)
     {
-        dump(' -- resource -- '.$resource);
         return 'api_resource' === $type;
     }
 
@@ -90,7 +82,7 @@ final class ResourceLoader extends Loader
      *
      * @throws RuntimeException
      */
-    private function addRoute(RouteCollection $routeCollection, string $resourceClass, string $operationName, array $operation, ResourceMetadata $resourceMetadata, string $operationType): void
+    private function addRoute(RouteCollection $collection, string $resourceClass, string $operationName, array $operation, ResourceMetadata $resourceMetadata, string $operationType): void
     {
         $resourceShortName = $resourceMetadata->getShortName();
 
@@ -98,7 +90,6 @@ final class ResourceLoader extends Loader
             if (!isset($operation['method'])) {
                 @trigger_error(sprintf('Not setting the "method" attribute is deprecated and will not be supported anymore in API Platform 3.0, set it for the %s operation "%s" of the class "%s".', OperationType::COLLECTION === $operationType ? 'collection' : 'item', $operationName, $resourceClass), E_USER_DEPRECATED);
             }
-
             return;
         }
 
@@ -106,16 +97,29 @@ final class ResourceLoader extends Loader
             throw new RuntimeException(sprintf('Either a "route_name" or a "method" operation attribute must exist for the operation "%s" of the resource "%s".', $operationName, $resourceClass));
         }
 
-        if (null === $controller = $operation['controller'] ?? null) {
-            $controller = sprintf('%s%s_%s', self::DEFAULT_ACTION_PATTERN, strtolower($operation['method']), $operationType);
-
+        $controller = $operation['controller'] ?? null;
+        if (null === $controller) {
+            $controller = sprintf(
+                '%s%s_%s',
+                self::DEFAULT_ACTION_PATTERN,
+                strtolower($operation['method']),
+                $operationType
+            );
             if (!$this->container->has($controller)) {
                 throw new RuntimeException(sprintf('There is no builtin action for the %s %s operation. You need to define the controller yourself.', $operationType, $operation['method']));
             }
         }
 
-        $path = trim(trim($resourceMetadata->getAttribute('route_prefix', '')), '/');
-        $path .= $this->operationPathResolver->resolveOperationPath($resourceShortName, $operation, $operationType, $operationName);
+        $path = trim(
+            trim($resourceMetadata->getAttribute('route_prefix', '')),
+            '/'
+        );
+        $path .= $this->operationPathResolver->resolveOperationPath(
+            $resourceShortName,
+            $operation,
+            $operationType,
+            $operationName
+        );
 
         $route = new Route(
             $path,
@@ -133,6 +137,12 @@ final class ResourceLoader extends Loader
             $operation['condition'] ?? ''
         );
 
-        $routeCollection->add(RouteNameGenerator::generate($operationName, $resourceShortName, $operationType), $route);
+        $collection->add(
+            RouteNameGenerator::generate(
+                $operationName,
+                $resourceShortName,
+                $operationType),
+            $route
+        );
     }
 }
